Refactor base executor
